requirements.txt
PyGithub==2.4.0
aiofiles==23.2.1
anthropic==0.36.0
duckduckgo-search==6.2.13
fastapi==0.104.1
google-api-python-client==2.142.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0
groq==0.9.0
httpx>=0.27.0
jinja2==3.1.2
langchain-anthropic==0.1.22
langchain-community==0.2.12
langchain-groq==0.1.9
langchain-openai==0.1.23
langchain==0.2.14
notion-client==2.2.1
ollama==0.3.1
openai==1.43.0
psutil>=5.9.8,<6.1
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
pyinstaller==6.14.2
python-dotenv==1.0.1
requests==2.32.3
slack_sdk==3.33.1
supabase==2.18.1
tavily-python==0.5.0
uvicorn==0.24.0
langchain-core==0.2.35
python-multipart>=0.0.9
python-multipart>=0.0.9
